Checking Branch (#147)
* Created Collapseable posts

Need to fix postScreenshot so post is correct height in commentview

Need to clean uop solution

* Post Loads correctly into comments view

For some reason the cells arent collapsing back to their original state after viewing comments. It works for some but not for others...

* Notification segues to myPosts

* load post into mycomments view

dequeueReusableCell

* Cell not expanding if not first posot

* Banned some words

Added alert if user create post containing a bad word

* Matched backtext constraints to frontText constraints (#143)

* home button from abilities view (#144)

changed navbar response from scrolltotop() to exitAbilities() if the abilities background view is being shown

* Finally Working

Only problem is posts collapse when they are off the screen. This is the only way I was able to make sure posts are collapsed after switching filters

* Clicking notifications opens comment view

* Removing changes for testing notifications

* Fixed error in styling for viewing collapsed post

* Merged from master fixed some styling

Co-authored-by: Cody Swain <swain.cody@gmail.com>